moving out the printer-helpers from the nodes folder
