+++ src/nodes/AssemblyCall.js
@@ -1,4 +1,4 @@
-const printSeparatedList = require('./print-separated-list');
+const { printSeparatedList } = require('../common/printer-helpers');
 
 const AssemblyCall = {
   print: ({ node, path, print, options }) =>
