+++ src/nodes/FunctionDefinition.js
@@ -5,9 +5,11 @@ const {
   util: { getNextNonSpaceNonCommentCharacterIndex }
 } = require('prettier');
 
-const printSeparatedList = require('./print-separated-list');
-const printSeparatedItem = require('./print-separated-item');
-const printComments = require('./print-comments');
+const {
+  printComments,
+  printSeparatedItem,
+  printSeparatedList
+} = require('../common/printer-helpers');
 
 const functionName = (node, options) => {
   if (node.isConstructor && !node.name) return 'constructor';
