+++ src/nodes/UsingForDeclaration.js
@@ -4,7 +4,7 @@ const {
   }
 } = require('prettier');
 
-const printSeparatedList = require('./print-separated-list');
+const { printSeparatedList } = require('../common/printer-helpers');
 
 const UsingForDeclaration = {
   print: ({ node, path, print, options }) => [
